refactor(gateway): move the SGLang field list out of the OpenAI provider router
The HTTP proxy strips SGLang-only request fields before forwarding, and it
imported the field list and the two strip helpers from the OpenAI
provider router. That made core proxying depend on a provider module,
which blocks building the gateway without provider routers.

Move the list and helpers, with their tests, into
`routers::common::sglang_fields`. The provider transformers and the HTTP
proxy now import from there. No behavior change.
